+++ admin/donate.php
@@ -3,7 +3,7 @@
 /**
  * The admin-facing functionality of the plugin.
  *
- * @package    Ultimate Facebook Comments
+ * @package    Ultimate Social Comments
  * @subpackage Admin
  * @author     Sayan Datta
  * @license    http://www.gnu.org/licenses/ GNU General Public License
@@ -25,10 +25,11 @@ function ufc_donate_admin_notice() {
     $no_thanks = wp_nonce_url( add_query_arg( 'ufc_donate_notice_action', 'ufc_no_thanks_donate_true' ), 'ufc_no_thanks_donate_true' ); ?>
     
     <div class="notice notice-success">
-        <p><?php _e( 'Hey, I noticed you\'ve been using Ultimate Facebook Comments for more than 2 week – that’s awesome! If you like Ultimate Facebook Comments and you are satisfied with the plugin, isn’t that worth a coffee or two? Please consider donating. Any amount is appreciated. Donations help me to continue support and development of this free plugin! Thank you very much!', 'ultimate-facebook-comments' ); ?></p>
+        <p><?php _e( 'Hey, I noticed you\'ve been using Ultimate Social Comments for more than 2 week – that’s awesome! If you like Ultimate Social Comments and you are satisfied with the plugin, isn’t that worth a coffee or two? Please consider donating. Any amount is appreciated. Donations help me to continue support and development of this free plugin! Thank you very much!', 'ultimate-facebook-comments' ); ?></p>
         <p><a href="https://www.paypal.me/iamsayan" target="_blank" class="button button-secondary"><?php _e( 'Donate Now', 'ultimate-facebook-comments' ); ?></a>&nbsp;
         <a href="<?php echo $dismiss; ?>" class="already-did"><strong><?php _e( 'I already donated', 'ultimate-facebook-comments' ); ?></strong></a>&nbsp;<strong>|</strong>
-        <a href="<?php echo $no_thanks; ?>" class="later"><strong><?php _e( 'Nope&#44; maybe later', 'ultimate-facebook-comments' ); ?></strong></a></p>
+        <a href="<?php echo $no_thanks; ?>" class="later"><strong><?php _e( 'Nope&#44; maybe later', 'ultimate-facebook-comments' ); ?></strong></a>&nbsp;<strong>|</strong>
+        <a href="<?php echo $dismiss; ?>" class="dismiss"><strong><?php _e( 'I don\'t want to donate', 'ultimate-facebook-comments' ); ?></strong></a></p>
     </div>
 <?php
 }