+++ public/templates/comments-template-empty.php
@@ -2,8 +2,7 @@
 /**
  * The public-facing functionality of the plugin.
  *
- * @category   HTML
- * @package    Ultimate WordPress Comments
+ * @package    Ultimate Social Comments
  * @subpackage Public
  * @author     Sayan Datta
  * @license    http://www.gnu.org/licenses/ GNU General Public License